refactor: pin-only Slack agent support
