+++ cli/templates/slack-manifest.json
@@ -65,8 +65,7 @@
         "reaction_added",
         "reaction_removed",
         "assistant_thread_started",
-        "assistant_thread_context_changed",
-        "app_context_changed"
+        "assistant_thread_context_changed"
       ]
     },
     "interactivity": {
