+++ src/slack/index.ts
@@ -13,7 +13,6 @@ import { createConversationSerializer } from "./conversation-view.ts";
 import { createApprovals } from "./approvals.ts";
 import { createTurnHandler } from "./turn-handler.ts";
 import { registerSlackEvents } from "./events.ts";
-import { createAgentPane } from "./agent-pane.ts";
 import { createSurfaceContextFulfiller } from "./surface-context.ts";
 import { createDeliveryPoller } from "./deliveries.ts";
 import { createDeferredAckReceiver } from "./deferred-ack.ts";
@@ -193,7 +192,6 @@ export async function startSlackPlugin(
     directory,
     ids,
     deduper,
-    agentPane: createAgentPane(),
     ...(cfg.webUiPublicUrl ? { webUiPublicUrl: cfg.webUiPublicUrl } : {}),
     ensureHeader,
   });
