+++ src/slack/turn-handler.ts
@@ -73,7 +73,6 @@ interface Incoming {
   files: SlackFile[];
   threadTs?: string;
   ts: string;
-  contextNote?: string;
   unprompted?: boolean;
   botAuthored?: boolean;
   synthetic?: boolean;
@@ -277,11 +276,7 @@ export function createTurnHandler(deps: {
       inc.kind === "dm"
         ? {
             location: "a direct message with the user",
-            details: {
-              channel: inc.channel,
-              ...(inc.threadTs ? { thread_ts: inc.threadTs } : {}),
-              ...(inc.contextNote ? { currently_viewing: inc.contextNote } : {}),
-            },
+            details: { channel: inc.channel, ...(inc.threadTs ? { thread_ts: inc.threadTs } : {}) },
             instructions: slackSurfaceInstructions(inc.kind),
             reactionGuidance: REACTION_DETECT_GUIDANCE,
             ...(ids.botHandle ? { botHandle: ids.botHandle } : {}),
